requirements.txt
imap_tools==0.38.1
python-dotenv==0.15.0
tinydb==4.4.0
